Crate datamodel_renderer

A set of datastructures meant for rendering a Prisma data model as a string. We don’t even try to make the result pretty. Please use the functionality of prisma-fmt for that.

All structs implement std::fmt::Display for easy usage.

An example use case to render a datasource as a string:

use datamodel_renderer::{Datasource, Env};
use indoc::{indoc, formatdoc};

let datasource = Datasource::new(
    "db",
    "postgres",
    Env::variable("DATABASE_URL")
);

// We get a string rendering without proper formatting
// by calling the `to_string()` method:
let rendered = datasource.to_string();

// The output is not formatted, so we call the reformat
// function to the result to make it look more kosher.
let rendered = psl::reformat(&rendered, 2).unwrap();

let expected = indoc! {r#"
    datasource db {
      provider = "postgres"
      url      = env("DATABASE_URL")
    }

"#};

assert_eq!(expected, &rendered);

// Additionally we can just pass the datasource to any
// format block to include it in the resulting string:
let rendered = formatdoc!(r#"
    {datasource}

    model A {{
      id Int @id
    }}
"#);

// Again, making the result indentation and spacing to
// look prettier.
let rendered = psl::reformat(&rendered, 2).unwrap();

let expected = indoc! {r#"
    datasource db {
      provider = "postgres"
      url      = env("DATABASE_URL")
    }

    model A {
      id Int @id
    }

"#};

assert_eq!(expected, &rendered);

Structs

Array

An array of values.

Configuration

The configuration part of a data model. First the generators, then the datasources.

Constant

A constant value. Should be used if a value has certain naming standards.

Datamodel

The PSL data model declaration.

Datasource

The datasource block in a PSL file.

Documentation

A documentation block on top of an item in the PSL.

Enum

An enum block in a PSL file.

EnumVariant

A variant declaration in an enum block.

Function

Represents a function value in the PSL.

Generator

The generator block of the datasource.

Text

Represents a string value in the PSL.

Enums

ConstantNameValidationError

Thrown if a constant cannot be cleanly created with the given input value.

Env

A value that can optionally be fetched from an environment variable.

FunctionParam

Represents a function parameter in the PSL.

Value

A PSL value representation.