Struct datamodel_renderer::configuration::Datasource

pub struct Datasource<'a> { /* private fields */ }

The datasource block in a PSL file.

Implementations

impl<'a> Datasource<'a>

pub fn new(name: &'a str, provider: &'a str, url: impl Into<Env<'a>>) -> Self

Create a new datasource with all required properties.

datasource db {
//         ^^ name
  provider = "postgresql"
//            ^^^^^^^^^^ provider
  url      = env("DATABASE_URL")
//                ^^^^^^^^^^^^ url
}

pub fn shadow_database_url(&mut self, url: impl Into<Env<'a>>)

Used for migration engine to reflect the contents of migrations directory to diff against the actual database.

datasource db {
  provider          = "postgresql"
  url               = env("DATABASE_URL")
  shadowDatabaseUrl = env("SHADOW_DATABASE_URL")
//                         ^^^^^^^^^^^^^^^^^^^ this
}

pub fn relation_mode(&mut self, relation_mode: RelationMode)

Who handles referential integrity.

datasource db {
  provider     = "postgresql"
  url          = env("DATABASE_URL")
  relationMode = "foreignKeys"
//                ^^^^^^^^^^^ this
}

pub fn push_custom_property(&mut self, key: &'a str, value: impl Into<Value<'a>>)

Add a custom connector-specific property to the datasource. Use this for settings that are only for a single database. For shared things, add an explicit property to the Datasource struct.

An example for PostgreSQL extensions, using an array of functions:

datasource db {
  provider   = "postgresql"
  url          = env("DATABASE_URL")
  extensions = [citext(version: "2.1")]
//             ^^^^^^^^^^^^^^^^^^^^^^^^ value
//^^^^^^^^^^ key
}

pub fn documentation(&mut self, documentation: impl Into<Cow<'a, str>>)

The documentation on top of the datasource.

/// This here is the documentation.
datasource db {
  provider = "postgresql"
  url      = env("DATABASE_URL")
}

pub fn from_psl(
psl_ds: &'a Datasource,
force_namespaces: Option<&'a [String]>
) -> Self

Create a rendering from a PSL datasource.