Module datamodel_renderer::datamodel

Types related to the datamodel section in the PSL.

Includes the model, enum and type definitions.

Structs

CompositeType

A type block in a PSL file.

CompositeTypeField

A field in a composite type block.

Datamodel

The PSL data model declaration.

DefaultValue

A field default value.

Enum

An enum block in a PSL file.

EnumVariant

A variant declaration in an enum block.

FieldType

A type of a field in the datamodel.

IdDefinition

Defines the id attribute in a model block.

IdFieldDefinition

Defines the id attribute in a field.

IndexDefinition

Defines an index in a model block.

IndexFieldInput

Input parameters for a field in a model index definition.

IndexFieldOptions

Options for a field-level index attribute.

Model

Defines a model block.

ModelField

A field in a model block.

Relation

Defines the relation argument of a model field.

Enums

IndexOps

Index type definition.