Module datamodel_renderer::datamodel

Types related to the datamodel section in the PSL.

Includes the model, enum and type definitions.

Structs

CompositeType

A type block in a PSL file.

Datamodel

The PSL data model declaration.

DefaultValue

A field default value.

Enum

An enum block in a PSL file.

EnumVariant

A variant declaration in an enum block.

Field

A field in a model block.

FieldType

A type of a field in the datamodel.

IdDefinition

Defines the id attribute in a model block.

IdFieldDefinition

Defines the id attribute in a field.

IndexDefinition

Defines an index in a model block.

IndexFieldInput

Input parameters for a field in a model index definition.

IndexOps

Index field operator definition.

Model

Defines a model block.

Relation

Defines the relation argument of a model field.

UniqueFieldAttribute

Options for a field-level unique attribute.

View

Defines a model block.